backend/pyproject.toml
[project]
name = "bpm-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "delawer33", email = "shukry.aliev@mail.ru"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.133.1,<0.134.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "alembic-postgresql-enum (>=1.10.0,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "argon2 (>=0.1.10,<0.2.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "boto3 (>=1.35.0,<2.0.0)",
    "pika (>=1.3.0,<2.0.0)",
    "aio-pika (>=9.4.0,<10.0.0)",
    "Pillow (>=11.0.0,<12.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "librosa (>=0.11.0,<0.12.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.4"

